Created Aggregator trait.
